Fixed exit error
If a user hit esc or closed the dialog menu in either of my menu
functions, it would still try to complete the function. So I fixed the
showConfirmDialog as well as a global int function. Finally I put in the
if loop to check if the user exited out, and let them naturally without
preforming any additional functions, and let them back in the main menu.